+++ tests/_lib.js
@@ -5,6 +5,7 @@ if (typeof process !== undefined && typeof module !== undefined) {
     nlp = require('../')
   } else {
     nlp = require('../src')
+    nlp.extend(require('../../plugins/strict-match'))
   }
 
   module.exports = nlp